Develop

+++ deploy/sql/challenge.sql
@@ -12,35 +12,35 @@ create sequence CHALLENGE_SEQ start with 1 increment by 50;
 create sequence RECIPE_SEQ start with 1 increment by 50;
 
 create table Challenge (
-    id bigint not null,
+    id int8 not null,
     authorID varchar(255),
     name varchar(255),
     primary key (id)
 );
-
+    
 create table Ingredient (
-    id bigint not null,
-    ingredientId bigint,
-    quantity integer not null,
-    CHALLENGE_ID bigint,
+    id int8 not null,
+    ingredientId int8,
+    quantity int4 not null,
+    CHALLENGE_ID int8,
     primary key (id)
 );
-
+    
 create table Recipe (
-    id bigint not null,
-    recipeId bigint,
-    CHALLENGE_ID bigint,
+    id int8 not null,
+    recipeId int8,
+    CHALLENGE_ID int8,
     primary key (id)
 );
 
-alter table Ingredient
-    add constraint FKaisyn8fyyxdnhm6knmgbt335i
-    foreign key (CHALLENGE_ID)
+alter table Ingredient 
+    add constraint FKaisyn8fyyxdnhm6knmgbt335i 
+    foreign key (CHALLENGE_ID) 
     references Challenge;
-
-alter table Recipe
-    add constraint FKk503phg1qn2raeqfsvghkwymq
-    foreign key (CHALLENGE_ID)
+    
+alter table Recipe 
+    add constraint FKk503phg1qn2raeqfsvghkwymq 
+    foreign key (CHALLENGE_ID) 
     references Challenge;
 
 INSERT INTO Challenge (authorID, name, id) values ('14', 'CREPESAMERE', nextval('Challenge_SEQ'));
